Gy_P_Admin

+++ src/main/java/com/bestseller/coffeestore/controller/AdminController.java
@@ -2,10 +2,12 @@ package com.bestseller.coffeestore.controller;
 
 import com.bestseller.coffeestore.controller.bean.DrinkCreation;
 import com.bestseller.coffeestore.controller.bean.ToppingCreation;
+import com.bestseller.coffeestore.dao.UsersDAO;
 import com.bestseller.coffeestore.dto.DrinkDTO;
 import com.bestseller.coffeestore.dto.MostUsedToppingDTO;
 import com.bestseller.coffeestore.dto.ToppingDTO;
 import com.bestseller.coffeestore.service.AdminService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -13,53 +15,74 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+@Slf4j
 @RestController
 @RequestMapping("/admin")
 public class AdminController {
 
     AdminService adminService;
+    UsersDAO usersDAO;
 
-    public AdminController(AdminService adminService) {
+    public AdminController(AdminService adminService, UsersDAO usersDAO) {
         this.adminService = adminService;
+        this.usersDAO = usersDAO;
     }
 
 //    CRUD endpoints for Drinks -----------------------------------------------------------
 
     @PostMapping(path = "/createdrink", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<DrinkDTO> createDrink(
+    public ResponseEntity<?> createDrink(
             @RequestBody DrinkCreation drinkCreation
     ) {
         if (drinkCreation == null) {
             // throw Exception (a custom exception could be created here...)
         }
-
-        return ResponseEntity.ok(adminService.createDrink(drinkCreation));
+        if (usersDAO.isAdmin(drinkCreation.getUserId())) {
+            return ResponseEntity.ok(adminService.createDrink(drinkCreation));
+        } else {
+            log.error("You are not authorized to create new drinks in the database!");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You are not authorized to create new drinks in the database!");
+        }
     }
 
-    @DeleteMapping(path = "/deletedrink/{drinkId}")
-    public ResponseEntity<Void> deleteDrink(
-            @PathVariable Long drinkId
+    @DeleteMapping(path = "/deletedrink/{drinkId}/{userId}")
+    public ResponseEntity<?> deleteDrink(
+            @PathVariable Long drinkId,
+            @PathVariable String userId
     ) {
         if (drinkId == null) {
             // throw Exception (a custom exception could be created here...)
         }
 
-        boolean isDeleted = adminService.deleteDrink(drinkId);
+        if (usersDAO.isAdmin(userId)) {
+            boolean isDeleted = adminService.deleteDrink(drinkId);
 
-        if (isDeleted) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if (isDeleted) {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            log.error("You are not authorized to delete drinks from the database!");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You are not authorized to delete drinks from the database!");
+        }
+
+    }
+
+    @GetMapping("/getAllDrinks/{userId}")
+    public ResponseEntity<?> getAllDrinks(
+            @PathVariable String userId
+    ) {
+        if (usersDAO.isAdmin(userId)) {
+            return ResponseEntity.ok(adminService.getAllDrinks());
+        } else {
+            log.error("You are not authorized to retrieve the list of drinks!");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You are not authorized to retrieve the list of drinks!");
         }
     }
 
-    @GetMapping("/getAllDrinks")
-    public ResponseEntity<List<DrinkDTO>> getAllDrinks() {
-        return ResponseEntity.ok(adminService.getAllDrinks());
-    }
-
     @PutMapping(value = "/updateDrink/{drinkId}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Void> updateDrink(
+    public ResponseEntity<?> updateDrink(
             @PathVariable Long drinkId,
             @RequestBody DrinkCreation drinkCreation
     ) {
@@ -68,53 +91,75 @@ public class AdminController {
             // throw Exception (a custom exception could be created here...)
         }
 
-        boolean isUpdated = adminService.updateDrink(drinkId, drinkCreation);
+        if (usersDAO.isAdmin(drinkCreation.getUserId())) {
+            boolean isUpdated = adminService.updateDrink(drinkId, drinkCreation);
 
-        if (isUpdated) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if (isUpdated) {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            log.error("You are not authorized to update drinks in the database!");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You are not authorized to update drinks in the database!");
         }
-
     }
 
 //    CRUD endpoints for Toppings -----------------------------------------------------------
 
     @PostMapping(path = "/createtopping", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<ToppingDTO> createTopping(
+    public ResponseEntity<?> createTopping(
             @RequestBody ToppingCreation toppingCreation
     ) {
         if (toppingCreation == null) {
             // throw Exception (a custom exception could be created here...)
         }
 
-        return ResponseEntity.ok(adminService.createTopping(toppingCreation));
+        if (usersDAO.isAdmin(toppingCreation.getUserId())) {
+            return ResponseEntity.ok(adminService.createTopping(toppingCreation));
+        } else {
+            log.error("You are not authorized to create new topping in the database!");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You are not authorized to create new topping in the database!");
+        }
     }
 
-    @DeleteMapping(path = "/deletetopping/{toppingId}")
-    public ResponseEntity<Void> deleteTopping(
-            @PathVariable Long toppingId
+    @DeleteMapping(path = "/deletetopping/{toppingId}/{userId}")
+    public ResponseEntity<?> deleteTopping(
+            @PathVariable Long toppingId,
+            @PathVariable String userId
     ) {
         if (toppingId == null) {
             // throw Exception (a custom exception could be created here...)
         }
 
-        boolean isDeleted = adminService.deleteTopping(toppingId);
+        if (usersDAO.isAdmin(userId)) {
+            boolean isDeleted = adminService.deleteTopping(toppingId);
 
-        if (isDeleted) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if (isDeleted) {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            log.error("You are not authorized to delete toppings from the database!");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You are not authorized to delete toppings from the database!");
         }
     }
 
-    @GetMapping("/getAllToppings")
-    public ResponseEntity<List<ToppingDTO>> getAllToppings() {
-        return ResponseEntity.ok(adminService.getAllToppings());
+    @GetMapping("/getAllToppings/{userId}")
+    public ResponseEntity<?> getAllToppings(
+            @PathVariable String userId
+    ) {
+        if (usersDAO.isAdmin(userId)) {
+            return ResponseEntity.ok(adminService.getAllToppings());
+        } else {
+            log.error("You are not authorized to retrieve the list of toppings!");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You are not authorized to retrieve the list of toppings!");
+        }
     }
 
     @PutMapping(value = "/updateTopping/{toppingId}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Void> updateTopping(
+    public ResponseEntity<?> updateTopping(
             @PathVariable Long toppingId,
             @RequestBody ToppingCreation toppingCreation
     ) {
@@ -123,19 +168,30 @@ public class AdminController {
             // throw Exception (a custom exception could be created here...)
         }
 
-        boolean isUpdated = adminService.updateTopping(toppingId, toppingCreation);
+        if (usersDAO.isAdmin(toppingCreation.getUserId())) {
+            boolean isUpdated = adminService.updateTopping(toppingId, toppingCreation);
 
-        if (isUpdated) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if (isUpdated) {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            log.error("You are not authorized to update toppings!");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You are not authorized to update toppings!");
         }
-
     }
 
-    @GetMapping("/mostUsedToppings")
-    public ResponseEntity<List<MostUsedToppingDTO>> getMostUsedToppings() {
-        return ResponseEntity.ok(adminService.getMostUsedToppings());
+    @GetMapping("/mostUsedToppings/{userId}")
+    public ResponseEntity<?> getMostUsedToppings(
+            @PathVariable String userId
+    ) {
+        if (usersDAO.isAdmin(userId)) {
+            return ResponseEntity.ok(adminService.getMostUsedToppings());
+        } else {
+            log.error("You are not authorized to get the list of the most used toppings!");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You are not authorized to get the list of the most used toppings!");
+        }
     }
 
 }
